requirements.txt
boto==2.42.0
dj-database-url==0.4.0
Django==1.9.2
django-cors-headers==1.1.0
django-filter==0.14.0
django-storages==1.5.0
djangorestframework==3.4.6
gunicorn==19.4.5
Markdown==2.6.6
Pillow==3.3.1
psycopg2==2.6.1
whitenoise==2.0.6
